Work toward supporting full data generator spec.
Previously restricted to data generators with a single variable and assuming math block was a simple identity. This work starts supporting the case when there are multiple variables and parameters used in a data generator. Has led to realisation that more work is needed to maybe separate variables from data sets. Currently this is building, by removing the actual serialisation of the data as that requires the math to now be interpreted.