starting work to support referencing models and making changes.
Need to be able to define a model once and then reuse it with different changes so removing the current
limitation on models always having a source that is the actual XML document URL and allowing the source
to be a reference to another model. Have not yet implemented support for any changes to the model.